+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 ### Latest changes
 
 
+* VFS-5879 Changed provider permissions to support permissions in space
 * VFS-5813 Added translations for new possible backend errors
 * VFS-1891 Added privacy policy and cookie consent notification
 * VFS-5769 Fixed render failure when switching between clusters